Ligatures only (#1)
* Add ligature support to Node engine

Introduces ligature support to the Node engine by updating `svgicons2svgfont` to v8.0.0.

* Add "ligaturesOnly" option to generate slimmer ligature fonts

The `ligatures` option results in an approximate doubling of output file sizes, as glyphs are repeated.  The new `ligaturesOnly` option removes the regular glyphs, so that the generated files are similar to those of a non-ligature font.